Attacking Interface and Distance Check
Now you can make a function called attack and it will automaticaly call the correct function for that soldier type by changing the Enum in the inspector.
Also it will check the distance between its closest target and its self with a function I made myself which is about 2x as fast.